[FIX] website_livechat: fix message reaction tour
The `test_complete_chatbot_flow_ui` ensures it is possible to add
reactions to the chat bot messages once the thread is saved on the
server.

This test was sometimes failing because the condition that is used to
wait for the thread to be persisted is incorrect: it checks that a
message with the user response exists but the chat bot question also
contains the response text.

This PR fixes the issue by using the `contains` helper instead: it
will wait until the expected state is reached.

fixes runbot-55291

closes odoo/odoo#151963

X-original-commit: 98c006803357a1dd1988bff3efff572b81d2e686
Signed-off-by: Didier Debondt (did) <did@odoo.com>
Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>